changes to seeach controller
