+++ search-app/app/Http/Controllers/SearchController.php
@@ -34,6 +34,23 @@ class SearchController extends Controller
             'data' => $validated,
         ]);
     }
+    public function runXQuery(Request $request)
+    {
+        // Log or process the received data
+        $data = $request->all(); // Get all incoming request data
+        \Log::info('Received data:', $data);
+        
+        // Placeholder function call
+        $response = $this->sayHello();
+
+        // Return the response
+        return response()->json(['message' => $response]);
+    }
+
+    private function sayHello()
+    {
+        return "We love you fariha";
+    }
 }
