+++ search-app/routes/web.php
@@ -2,7 +2,7 @@
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\RouteController;
-use App\Http\Controllers\XmlController;
+use App\Http\Controllers\SearchController;
 
 
 //Default route for the home page
@@ -11,4 +11,4 @@ Route::get('/', [RouteController::class, 'display_view'])->defaults('page', 'hom
 //route when user goes to a page
 Route::get('/{page}', [RouteController::class, 'display_view']);
 
-Route::post('/xquery', [XmlController::class, 'runXQuery']);
+Route::post('/xquery', [SearchController::class, 'runXQuery']);
